Add some unit tests using GTest and Catch2

+++ tests/unit/algebra/linear_equation_system.cpp
@@ -6,8 +6,8 @@ namespace raytracing::Algebra {
 
 TEST(Matrix3x3DoubleTest, UniqueSolution) {
 	// given
-	Eigen::Matrix3d A{{1, 6, 2}, {1, 3, 1}, {1, 2, 3}}; // NOLINT
-	Eigen::Vector3d b{1, 2, 3};
+	Eigen::Matrix3f A{{1, 6, 2}, {1, 3, 1}, {1, 2, 3}}; // NOLINT
+	Eigen::Vector3f b{1, 2, 3};
 
 	// when
 	auto solution = eq_sys_solve(A, b);
@@ -18,8 +18,8 @@ TEST(Matrix3x3DoubleTest, UniqueSolution) {
 
 TEST(Matrix3x3DoubleTest, NoSolution) {
 	// given
-	Eigen::Matrix3d A{{1, 2, 3}, {4, 5, 6}, {7, 8, 9}}; // NOLINT
-	Eigen::Vector3d b{3, 3, 4};
+	Eigen::Matrix3f A{{1, 2, 3}, {4, 5, 6}, {7, 8, 9}}; // NOLINT
+	Eigen::Vector3f b{3, 3, 4};
 
 	// when
 	auto solution = eq_sys_solve(A, b);
@@ -31,8 +31,8 @@ TEST(Matrix3x3DoubleTest, NoSolution) {
 
 TEST(Matrix3x3DoubleTest, InfiniteSolution) {
 	// given
-	Eigen::Matrix3d A{{1, 2, 3}, {2, 4, 6}, {3, 6, 9}}; // NOLINT
-	Eigen::Vector3d b{1, 2, 3};
+	Eigen::Matrix3f A{{1, 2, 3}, {2, 4, 6}, {3, 6, 9}}; // NOLINT
+	Eigen::Vector3f b{1, 2, 3};
 
 	// when
 	auto solution = eq_sys_solve(A, b);
